Pass any initial :jq options that begin with "-" as separate words.
Previously the jq-header processing would call `shell-quote-argument` on
the entire jq argument string so that it would treated as the jq filter,
but that meant that it was not possible to pass command-line options to jq
itself (such as `--compact-output` or `--sort-keys`).  Now any initial
words are passed separately, without being quoted, to make that possible.